Add modules. Reactor results into sections
