+++ emulation_system/tests/e2e/fixtures/ot3_containers.py
@@ -8,7 +8,7 @@ from docker.models.containers import Container  # type: ignore[import]
 from emulation_system.compose_file_creator.config_file_settings import (
     RepoToBuildArgMapping,
 )
-from tests.e2e.utilities.build_arg_configurations import BuildArgConfigurations
+from tests.e2e.test_definition.build_arg_configurations import BuildArgConfigurations
 from tests.e2e.utilities.helper_functions import get_mounts
